[Fix] API manager 추가 - 기존 exteion에 있던 것을, 공통적으로 사용하기 위하여 Singleton pattern으로 새로 구성함
